Organize files of bipartite features
