+++ lib/main/graph/bipartite/main-bipartite-check.cpp
@@ -1,24 +1,24 @@
 #include <iostream>
 #include "template/template-main.hpp"
-#include "cpalgo/graph/bipartite-maximum-matching.hpp"
+#include "cpalgo/graph/bipartite/bipartite_check.hpp"
 
 using namespace std;
 
-BipartiteMaximumMatching solver;
+BipartiteCheck solver;
 
 void action_init() {
     size_t size;
     cin >> size;
-    solver = BipartiteMaximumMatching(size);
+    solver = BipartiteCheck(size);
     cout << "true" << endl;
 }
 
-void action_edge() {
+void action_add_edge() {
     size_t u, v;
     cin >> u >> v;
     if (u >= solver.size() ||
-        v >= solver.size())
-    {
+        v >= solver.size()
+    ) {
         cout << "false" << endl;
         return;
     }
@@ -27,17 +27,13 @@ void action_edge() {
 }
 
 void action_solve() {
-    vector<size_t> match;
-    size_t ans = solver.solve(match);
-    cout << ans << endl;
-    for (size_t i = 0; i < match.size(); ++i) {
-        cout << i << ": " << match[i] << endl;
-    }
+    bool ans = solver.solve();
+    cout << (ans ? "true" : "false") << endl;
 }
 
 void setup(string& header, map<string,Command>& commands) {
-    header = "Bipartite Maximum Matching";
+    header = "Bipartite Check";
     commands["init"] = { "init {size}", action_init };
-    commands["edge"] = { "edge {u} {v}", action_edge };
+    commands["edge"] = { "edge {u} {v}", action_add_edge };
     commands["solve"] = { "solve", action_solve };
 }
