+++ lib/main/graph/bipartite/main-bipartite-maximum-matching.cpp
@@ -1,24 +1,24 @@
 #include <iostream>
 #include "template/template-main.hpp"
-#include "cpalgo/graph/bipartite-check.hpp"
+#include "cpalgo/graph/bipartite/bipartite_maximum_matching.hpp"
 
 using namespace std;
 
-BipartiteCheck solver;
+BipartiteMaximumMatching solver;
 
 void action_init() {
     size_t size;
     cin >> size;
-    solver = BipartiteCheck(size);
+    solver = BipartiteMaximumMatching(size);
     cout << "true" << endl;
 }
 
-void action_add_edge() {
+void action_edge() {
     size_t u, v;
     cin >> u >> v;
     if (u >= solver.size() ||
-        v >= solver.size()
-    ) {
+        v >= solver.size())
+    {
         cout << "false" << endl;
         return;
     }
@@ -27,13 +27,17 @@ void action_add_edge() {
 }
 
 void action_solve() {
-    bool ans = solver.solve();
-    cout << (ans ? "true" : "false") << endl;
+    vector<size_t> match;
+    size_t ans = solver.solve(match);
+    cout << ans << endl;
+    for (size_t i = 0; i < match.size(); ++i) {
+        cout << i << ": " << match[i] << endl;
+    }
 }
 
 void setup(string& header, map<string,Command>& commands) {
-    header = "Bipartite Check";
+    header = "Bipartite Maximum Matching";
     commands["init"] = { "init {size}", action_init };
-    commands["edge"] = { "edge {u} {v}", action_add_edge };
+    commands["edge"] = { "edge {u} {v}", action_edge };
     commands["solve"] = { "solve", action_solve };
 }
